+++ src/commonMain/kotlin/ui/Settings.kt
@@ -5,6 +5,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
 
 class Settings {
+    val runOnStartup = mutableStateOf(false)
     val showAllText = mutableStateOf(false)
     val reallyShowAllText = mutableStateOf(false)
     val enableDiscordEmbed = mutableStateOf(true)